feat(search): unify search behind search worker
