+++ .github/workflows/index-semantic-podcasts.yml
@@ -47,7 +47,7 @@ jobs:
           SIMPLECAST_KEY: ${{ secrets.SIMPLECAST_KEY }}
           CHATS_WITH_KENT_PODCAST_ID: ${{ secrets.CHATS_WITH_KENT_PODCAST_ID }}
 
-  sync-lexical-search:
+  sync-search-worker:
     needs: index
-    uses: ./.github/workflows/sync-lexical-search-worker.yml
+    uses: ./.github/workflows/sync-search-worker.yml
     secrets: inherit
