+++ other/compute-deploy-plan.ts
@@ -15,13 +15,13 @@ const defaultFetchTimeoutMs = 10_000
 const nonFlyDeployablePathPrefixes = [
 	'services/site/content/',
 	'services/call-kent-audio-worker/',
-	'services/lexical-search-worker/',
+	'services/search-worker/',
 	'services/oauth/',
 ]
 
 const nonFlyDeployableFiles = new Set([
 	'.github/workflows/deploy-call-kent-audio-worker.yml',
-	'.github/workflows/deploy-lexical-search-worker.yml',
+	'.github/workflows/deploy-search-worker.yml',
 	'.github/workflows/deploy-oauth-worker.yml',
 ])
 
@@ -38,10 +38,8 @@ const callKentAudioWorkerFiles = new Set([
 	'.github/workflows/deploy-call-kent-audio-worker.yml',
 ])
 
-const lexicalSearchWorkerPathPrefixes = ['services/lexical-search-worker/']
-const lexicalSearchWorkerFiles = new Set([
-	'.github/workflows/deploy-lexical-search-worker.yml',
-])
+const searchWorkerPathPrefixes = ['services/search-worker/']
+const searchWorkerFiles = new Set(['.github/workflows/deploy-search-worker.yml'])
 
 const oauthWorkerPathPrefixes = ['services/oauth/']
 const oauthWorkerFiles = new Set(['.github/workflows/deploy-oauth-worker.yml'])
@@ -62,8 +60,8 @@ const DEPLOY_ENVIRONMENTS: Record<string, (refName: string) => string | null> =
 			refName === 'main' ? 'oauth-production' : null,
 		deployCallKentAudioWorker: (refName) =>
 			refName === 'main' ? 'call-kent-audio-worker-production' : null,
-		deployLexicalSearchWorker: (refName) =>
-			refName === 'main' ? 'lexical-search-worker-production' : null,
+		deploySearchWorker: (refName) =>
+			refName === 'main' ? 'search-worker-production' : null,
 	}
 
 type ChangedFile = {
@@ -389,7 +387,7 @@ export async function computeDeployPlan({
 		pushChangedFiles,
 		refreshResult,
 		siteDeployResult,
-		lexicalWorkerDeployResult,
+		searchWorkerDeployResult,
 		oauthDeployResult,
 		audioWorkerDeployResult,
 	] = await Promise.all([
@@ -424,7 +422,7 @@ export async function computeDeployPlan({
 			? getDeploymentChangedFiles({
 					currentCommitSha,
 					refName,
-					target: 'deployLexicalSearchWorker',
+					target: 'deploySearchWorker',
 					owner: repo.owner,
 					repo: repo.repo,
 					token,
@@ -479,10 +477,10 @@ export async function computeDeployPlan({
 			pathPrefixes: semanticContentPathPrefixes,
 			runWhenUnknown: isPushEvent,
 		}),
-		deployLexicalSearchWorker: shouldRunPathTarget({
-			changedFiles: lexicalWorkerDeployResult.changedFiles,
-			pathPrefixes: lexicalSearchWorkerPathPrefixes,
-			files: lexicalSearchWorkerFiles,
+		deploySearchWorker: shouldRunPathTarget({
+			changedFiles: searchWorkerDeployResult.changedFiles,
+			pathPrefixes: searchWorkerPathPrefixes,
+			files: searchWorkerFiles,
 			runWhenUnknown: isPushEvent,
 		}),
 		deployCallKentAudioWorker: shouldRunPathTarget({
@@ -531,9 +529,7 @@ export async function writeDeployPlanOutputs({
 		`deploy_site=${String(deployPlan.deploySite)}`,
 		`refresh_content=${String(deployPlan.refreshContent)}`,
 		`index_semantic_content=${String(deployPlan.indexSemanticContent)}`,
-		`deploy_lexical_search_worker=${String(
-			deployPlan.deployLexicalSearchWorker,
-		)}`,
+		`deploy_search_worker=${String(deployPlan.deploySearchWorker)}`,
 		`deploy_call_kent_audio_worker=${String(
 			deployPlan.deployCallKentAudioWorker,
 		)}`,
