--- services/lexical-search-worker/src/env.d.ts
@@ -1,8 +1,0 @@
-export type Env = {
-	LEXICAL_SEARCH_DB: D1Database
-	LEXICAL_SEARCH_WORKER_TOKEN: string
-	R2_ENDPOINT: string
-	R2_BUCKET: string
-	R2_ACCESS_KEY_ID: string
-	R2_SECRET_ACCESS_KEY: string
-}
