+++ services/search-worker/src/env.d.ts
@@ -1,0 +1,11 @@
+export type Env = {
+	SEARCH_DB: D1Database
+	SEARCH_INDEX: VectorizeIndex
+	SEARCH_ARTIFACTS_BUCKET: R2Bucket
+	SEARCH_WORKER_TOKEN: string
+	CLOUDFLARE_ACCOUNT_ID: string
+	CLOUDFLARE_API_TOKEN: string
+	CLOUDFLARE_AI_EMBEDDING_GATEWAY_ID: string
+	CLOUDFLARE_AI_GATEWAY_AUTH_TOKEN: string
+	CLOUDFLARE_AI_EMBEDDING_MODEL: string
+}
