+++ services/search-worker/tsconfig.json
