+++ services/site/.env.example
@@ -98,12 +98,12 @@ VERIFIER_API_KEY=some_api_key
 # Feature: Cloudflare OAuth
 CF_INTERNAL_SECRET=some_long_random_string
 
-# Feature: lexical search worker
+# Feature: /search
 # Mocked: yes (site calls a mock Worker URL intercepted by MSW in dev/test)
-LEXICAL_SEARCH_WORKER_URL=https://mock.lexical-search-worker.local
-LEXICAL_SEARCH_WORKER_TOKEN=local-dev-lexical-search-token
+SEARCH_WORKER_URL=https://mock.search-worker.local
+SEARCH_WORKER_TOKEN=local-dev-search-token
 
-# Feature: /search (semantic search via Cloudflare Workers AI + Vectorize)
+# Feature: search indexing + semantic-search admin
 # Mocked: yes when CLOUDFLARE_API_TOKEN starts with MOCK (otherwise passes through)
 CLOUDFLARE_ACCOUNT_ID=MOCK_CLOUDFLARE_ACCOUNT_ID
 CLOUDFLARE_API_TOKEN=MOCK_CLOUDFLARE_API_TOKEN
