+++ services/site/app/utils/vectorize.server.ts
@@ -1,0 +1,57 @@
+import { getEnv } from '#app/utils/env.server.ts'
+
+function getCloudflareApiBaseUrl() {
+	return 'https://api.cloudflare.com/client/v4'
+}
+
+function getRequiredVectorizeEnv() {
+	const env = getEnv()
+	return {
+		accountId: env.CLOUDFLARE_ACCOUNT_ID,
+		apiToken: env.CLOUDFLARE_API_TOKEN,
+		indexName: env.CLOUDFLARE_VECTORIZE_INDEX,
+	}
+}
+
+export async function vectorizeDeleteByIds({
+	ids,
+}: {
+	ids: string[]
+}): Promise<unknown> {
+	const { accountId, apiToken, indexName } = getRequiredVectorizeEnv()
+	if (!Array.isArray(ids) || ids.length === 0) {
+		return { result: { deleted: 0 } }
+	}
+
+	const body = JSON.stringify({ ids })
+	const doFetch = (path: string) => {
+		const url = `${getCloudflareApiBaseUrl()}/accounts/${accountId}${path}`
+		return fetch(url, {
+			method: 'POST',
+			headers: {
+				Authorization: `Bearer ${apiToken}`,
+				'Content-Type': 'application/json',
+			},
+			body,
+		})
+	}
+	const throwIfNotOk = async (response: Response, path: string) => {
+		if (response.ok) return
+		const text = await response.text().catch(() => '')
+		throw new Error(
+			`Cloudflare API request failed: ${response.status} ${response.statusText} (${path})${text ? `\n${text}` : ''}`,
+		)
+	}
+
+	const v2Path = `/vectorize/v2/indexes/${indexName}/delete_by_ids`
+	const v2Response = await doFetch(v2Path)
+	if (v2Response.status === 404 || v2Response.status === 405) {
+		const legacyPath = `/vectorize/indexes/${indexName}/delete_by_ids`
+		const legacyResponse = await doFetch(legacyPath)
+		await throwIfNotOk(legacyResponse, legacyPath)
+		return (await legacyResponse.json()) as unknown
+	}
+
+	await throwIfNotOk(v2Response, v2Path)
+	return (await v2Response.json()) as unknown
+}
